sock/uring: Separate POLLIN from POLLERR tasks
The POLLIN task is going to be associated with a read operation
in an upcoming patch and cannot be bundled with the POLLERR.

The POLLIN task is only submitted if the socket does not already have
pending_recv set, but POLLERR is always submitted. This is a shift
toward our new model where the commands stay submitted always. However,
if a POLLERR completes successfully while the socket is being removed
from a group, then an errqueue could get submitted. So new flags are
added during removal to avoid submitting new commands.

Change-Id: Ib1bb88cb0fa7c5e7237c9e99679bdcb4b4a7e3c7
Signed-off-by: Ben Walker <benjamin.walker@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/17593
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>
Community-CI: Mellanox Build Bot