test/autobuild_common: modify native DPDK repository fetching
To reduce the number of requests to clone an external DPDK
repository, check if there is already a DPDK repository at the
given location. Currently, in our CI, we use tools to fetch repositories
and we can pass their location in a parameter. If no repository is
available at the specified location, as a fall back, clone it locally.

Signed-off-by: Kamil Godzwon <kamilx.godzwon@intel.com>
Change-Id: I999b71e6f2200e8fbac07b543ed11212b5e5c09b
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/17553
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>