pkgdep/git: Set bcc's git refspec
When bcc is handled as a bpftrace's submodule, it's not able to
detect its own revision as it depends on .git/HEAD which does
not exist in such a case. This taints the build log with warn
similar to:

CMake Warning at cmake/version.cmake:31 (message):
  Could not extract major/minor/patch from revision EAD-HASH-NOTFOUND-
Call Stack (most recent call first):
  CMakeLists.txt:98 (include)
-- Revision is EAD-HASH-NOTFOUND- (major , minor , patch )

which at very best is just confusing so make sure the log is clean.

Change-Id: Id215174f3187ca1180abcf75a6eeab799d74808b
Signed-off-by: Michal Berger <michal.berger@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/17636
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Kamil Godzwon <kamilx.godzwon@intel.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>