app: add auto-calculation of msg_mempool_size
Fixes #811

Currently the message memory pool size is always
configured to the same default value, regardless
of the number of cores the SPDK application is
running on.
In some situations, the default value of 262143
would not be sufficient, especially if large
number of SPDK's threads is started.
This patch adds calculation of msg_mempool_size, if
more than 63 cores are assigned to SPDK. In such case
the size of msg_mempool_size is calculated as
core_count * 4096.
For up to 63 cores, the value of 262143 is always used.

The user can also specify size of the message
memory pool with --msg-mempool-size option.
If specified byt the user, it always overrides
default or automatically calculated size.

Change-Id: Id5c57df4dd6492dbab6d56d49ec2ebe52ee7440d
Signed-off-by: Marcin Spiewak <marcin.spiewak@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/17831
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>