ublk: use temporary malloc buffer for kernel workaround
Some older kernels require a payload buffer to be attached
to a posted ublk io, even though the buffer isn't required
until later (either when committing a read I/O completion,
or posting a buffer in response to a NEED_DATA completion).

So instead of worrying about allocating a bunch of
valid I/O buffers for this case, just keep it simple and
use a single malloc buffer for the initial IO posting,
then immediately free that buffer.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: I59096cdb17fd6fa01adf67ffe54f2a0547a4fafb
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/17921
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Changpeng Liu <changpeng.liu@intel.com>