nvmf: use spdk_bit_array to track subsystem ids
Currently the subsystem id (sid) is just the index
of the subsystem pointer into the target's subsystems
pointer array.  But this array is going away in
upcoming patches, while we still need subsystem ids
for things like subsystem poll groups.

So transition to that end scenario by introducing
an spdk_bit_array to track the subsystem ids.  So
now we will look for a free bit to get a free
sid, and then put the subsystem pointer into that
index in the target's subsystems array.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: I0f01dfdec6f22c4937df63e50e05650a17c82fe1
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/17965
Reviewed-by: Ben Walker <benjamin.walker@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>