test/unit/bdev: remove bdev_multi_allocation test
This test is not really useful. The intended purposes
I think were:

1) Validate that the height of the tree is never too
   high.
2) Capture the time to insert a bdev's names into the
   rb tree.

But #1 isn't being tested correctly.  It only tests
the height of a name that was *just* inserted.  RB
trees always insert a new node at the leaf, followed
by some rotations as necessary to maintain the RB
tree properties.

#2 doesn't help with anything.  The NOTICELOGs don't
get printed, and even if they did, we aren't validating
them at all.

So just remove this test altogether.  We don't need
to validate the RB tree code itself, and we shouldn't
try to validate performance with a unit test.  The way
the test is constructed is way too expensive too - it
takes up the vast majority of the time to run the
entire unit test suite. Removing just this one test cut
the time for unittest.sh from 28 seconds to less than
3 seconds on my system.

Fixes issue #3013.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: If2253d4927ca5f1184fb165ed3f2c25befe9f4a4
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/17980
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>