test/vhost: Adjustments for bigger nvmes
E.g. when vhost tests work with a 8TB drive the bdev_lvol_delete()
needs a bit more time to complete, hence the default timeout must
be adjusted.

Also, in case we are working with *TB drives, the unit of the target
partitions can also change due to their size - instead of hard-coded
"G", use $(S) to match on any string (but keep the $(FP) as well to
indicate we are still looking for a floating number).

Change-Id: Ic11e71126071717846d53e3b152a37ae6521314b
Signed-off-by: Michal Berger <michal.berger@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/17988
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>
Reviewed-by: Kamil Godzwon <kamilx.godzwon@intel.com>
Reviewed-by: Karol Latecki <karol.latecki@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Jaroslaw Chachulski <jaroslawx.chachulski@intel.com>