test/nvme/sw_hotplug: Use timing_cmd to debug hotplug events
This is done to keep track of how much time it takes for the
sw_hotplug to finish its tests depending on how many drives
are attached to hotplug.

Due to the https://review.spdk.io/gerrit/c/spdk/spdk/+/18078 the
time to wait for each set of hotplug events is adjusted based on
the number of drives which means that each drive will add an extra
overhead - currently, this makes these tests to run for almost 2m
with just a single drive.

Debugging this process will help to adjust these tests in the future
if needed.

Change-Id: I2244ff769a85cf8407d8efe0e695ce5f124e146b
Signed-off-by: Michal Berger <michal.berger@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/18074
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Broadcom CI <spdk-ci.pdl@broadcom.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>