nvme/fio_plugin: fix check with blocksizes
As there are other ways of specifying block sizes, such as
bsrange or bssplit, the existing check which validates if
block size is either multiple of LBA size or extended LBA
is not true.
So, if bsrange is specified instead of bs, the td->o.bs[ddir]
values are initialized to default 4096. So for namespaces
formatted with extended LBA, the existing check doesn't work.
The correct check should be against td->o.min_bs[ddir] and
td->o.max_bs[ddir].

Note if only bs or blocksize option is specified, fio will
initialize min_bs and max_bs values to bs.

Change-Id: I4c546cae2c3b406851afe3edec0eed239c98e99e
Signed-off-by: Ankit Kumar <ankit.kumar@samsung.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/18060
Reviewed-by: Jim Harris <james.r.harris@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>