util: spdk_ioviter is now variable size
The size of the structure needed can be quickly calculated using
SPDK_IOVITER_SIZE. This size can be used to allocate on the stack or the
heap. All current uses allocate on the stack.

For backward compatibility, struct spdk_ioviter has a default size of
two and can be directly placed onto the stack without calling
SPDK_IOVITER_SIZE.

Change-Id: I19858f1241d44575ebba5bacae9c560fc9dc56c7
Signed-off-by: Ben Walker <benjamin.walker@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/18066
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Artur Paszkiewicz <artur.paszkiewicz@intel.com>