bdev/rbd: Share the rados_ioctx_t if images are from same pool
For rbd images belonging to the same rados pool,
we can share the rados_ioctx_t structure, then it
can save the librados related resource in client side.

Because there is a configuration parameter `librados_thread_count`
which is reponsible for io_context_pool. Then if
we can share the same rados_ioctx_t if we have the same rados
pool under shared rados_t. It can save the client resources when
there are multiple images in the same pool.

Change-Id: I7f71191d7dd72eec359840a5767da74ff6e4ec92
Signed-off-by: Ziye Yang <ziye.yang@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/18107
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>
Reviewed-by: Changpeng Liu <changpeng.liu@intel.com>