nvmf/tcp: Fix segfault in nvmf_tcp_poll_group_create
If any of goto labels in this function s triggered,
then nvmf_tcp_poll_group_destroy is called and last
function tries to remove the poll group from transport
groups list. But the poll group is not in the list at
this moment, so we have a segmentation fault.
To fix this issue, check if the poll group has an
assigned transport - that is done when
nvmf_tcp_poll_group_create completes. The same is
done in RDMA transport.

Signed-off-by: Alexey Marchuk <alexeymar@nvidia.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/18149 (master)

(cherry picked from commit e744ee8028347b9c222a9535a79022ae0f78dab8)
Change-Id: I0e3125524c199de6e6e4321913f9bf2f09a56712
Signed-off-by: Krzysztof Karas <krzysztof.karas@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/18379
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>