nvmf/tcp: Fix segfault in nvmf_tcp_poll_group_create
If any of goto labels in this function s triggered,
then nvmf_tcp_poll_group_destroy is called and last
function tries to remove the poll group from transport
groups list. But the poll group is not in the list at
this moment, so we have a segmentation fault.
To fix this issue, check if the poll group has an
assigned transport - that is done when
nvmf_tcp_poll_group_create completes. The same is
done in RDMA transport.

Signed-off-by: Alexey Marchuk <alexeymar@nvidia.com>
Change-Id: I0e3125524c199de6e6e4321913f9bf2f09a56712
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/18149
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <james.r.harris@intel.com>