test/unit/event: explicitly align reactor on stack
This seems to be a bug in Fedora 38 gcc - if struct
is declared with alignment, then it should get
allocated with alignment on stack.

But work around it for now by explicitly declaring
the variable as aligned.

Fixes issue #3004.

(cherry-picked from 771a7b8792a5c3cca134bfc8639479c9826b29cc)
Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: I52d6051815ea55c67f5c0a27648a108c232013e5
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/18214 (master)
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/20909
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>