bdev/nvme: nvme_ctrlr_op_rpc() can enable/disables ctrlr dynamically
Extend nvme_ctrlr_op_rpc() and nvme_bdev_ctrlr_op_rpc() to support enable
or disable a controller or all controllers in a bdev controller.

To disable a controller, bdev_nvme_disable_ctrlr() cancels reconnect
and disable a controller if reconnect is already scheduled, or disconnect
and disable a controller. Disable is to keep a controller disconnected
without scheduling a reconnect.

To enable a controller, bdev_nvme_enable_ctrlr() reconnects a controller
if it is disabled.

To indicate a controller is disabled, add a disabled variable to the
nvme_ctrlr structure.

Signed-off-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Change-Id: I02f97cdc549f317f4d37c802a125bf0f0db855fe
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/18235
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Richael <richael.zhuang@arm.com>
Reviewed-by: Ben Walker <benjamin.walker@intel.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>