build: regenerate modules pkgconfig files when config changes
If modules are added or removed from the current configuration
when re-running "configure" (i.e. --with-idxd), we need to
also generate the associated pkgconfig file that lists all of
the modules for that category (i.e. spdk_accel_modules.pc).

Do this by adding mk/config.mk as a dependency for each of the
modules pkgconfig files.

Fixes issue #3018.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: Ied37f378aeba2eddc44c0649710a9e8e43e4d35b
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/18233
Reviewed-by: Ben Walker <benjamin.walker@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>