nvme: Fix ZNS assert for NS_ATTR_CHANGED AEN
For a ZNS drive, there is a block of memory (nsdata_zns) that is
allocated when the ZNS drive is enumerated. This memory is already
allocated when the controller code goes to re-construct the
namespace for an SPDK_NVME_ASYNC_EVENT_NS_ATTR_CHANGED AEN so the
controller code asserts.

This commit will fix the assert by destroying (freeing) the
namespace structure before re-creating it.

Fixes #3028

Change-Id: I2e75077dce727cc1c912c8f980e4d78c1be328ba
Signed-off-by: Curt Bruns <curt.e.bruns@gmail.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/18251
Reviewed-by: Changpeng Liu <changpeng.liu@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <james.r.harris@intel.com>