test: Cleanup timing.txt
Remove extra time() calls which are not necessary for code blocks
which are executed under run_test().

Also, drop random timing_exit() - the "exit" flow is not testname-based,
hence it always affects current stack (i.e. it simply pops the last
test without caring if it matches the testname argument or not)
tainting the time accounting.

Case in point, for nvme test suit, this rogue timing_exit() call was
causing "autotest;lib" time to be < 0 under the nvme-phy-autotest job
since the "autotest;lib;nvme" was being put on the test stack twice
(so the total child time for "autotest;lib" was bigger then the actual
time that passed between "autotest;lib" push and pop).

Change-Id: Iaa491a8117cb656f81c0979ccd1fa98e275bb7f9
Signed-off-by: Michal Berger <michal.berger@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/18296
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Pawel Piatek <pawelx.piatek@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>