test/rocksdb: update commit id to spdk-v8.1.1 branch
The latest RocksDB also generates a librocksdb.a
file that far exceeds the artifact limit imposed by
SPDK CI.  But we don't need to save any of the build
objects from the rocksdb directory, so just do a
"git clean" at the end of the test, as part of moving
to the new spdk-v8.1.1 branch.

Fixes issue #2999.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: I0c7214077aedcdc49091026163c45e343672e36b

Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/18354
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Ben Walker <benjamin.walker@intel.com>
Reviewed-by: Kamil Godzwon <kamilx.godzwon@intel.com>
Reviewed-by: Michal Berger <michal.berger@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>