test/rocksdb: update commit id to spdk-v8.1.1 branch
The latest RocksDB also generates a librocksdb.a
file that far exceeds the artifact limit imposed by
SPDK CI.  But we don't need to save any of the build
objects from the rocksdb directory, so just do a
"git clean" at the end of the test, as part of moving
to the new spdk-v8.1.1 branch.

Fixes issue #2999.

(cherry-picked from 19ce0392de8615937480713f70c52c02806fb61b)
Change-Id: I0c7214077aedcdc49091026163c45e343672e36b
Signed-off-by: Jim Harris <james.r.harris@intel.com>

Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/18354 (master)
Change-Id: If2b7d2f23eed8a91e15e1943c5b12b327d3769f9
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/20892
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Reviewed-by: Karol Latecki <karol.latecki@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>