scripts/setup: Attempt to unbind all devices from target iommu group
Depending on the pci topology, there may be setups where target nvme
is part of iommu group which consists of more devices bound to
various different drivers. In such a case, DPDK will refuse to
configure the device and mark given iommu group as "not viable".

To workaround this, we may try to unbind all remaining devices from
their drivers (binding to vfio or uio may fail, depending on what
kind of device we are dealing with). However, since this is an
invasive operation it should be done only on demand.

Successfully tested under WFP platform where nvme was connected to
fanout pcie switches, however, results may vary across the systems.

Change-Id: I4f315cf2e4178841859ecbba5ca09cae5fcbd497
Signed-off-by: Michal Berger <michal.berger@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/18357
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Karol Latecki <karol.latecki@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>