test/nvme_perf: remove preconditioning
This script is designed in a way to call it for one
use case. Executing preconditioning each time we call
this script would be an overkill. Additionally, we
plan to add support for lvol bdevs, so we would need
to add special protection to avoid erase of lvol data
from disks by preconditioning.

Signed-off-by: Pawel Piatek <pawelx.piatek@intel.com>
Change-Id: I5c2dd457d995c2743bf2748dc835e65dbff7cdcc
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/18412
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Karol Latecki <karol.latecki@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Michal Berger <michal.berger@intel.com>