test/abi: match any LTS release for SO major bump check
After every LTS major SO is increased to allow minor SO
changes for period when particular LTS is supported.
Meanwhile ABI test verifies if version was increased
without modifying the actual implementation.
Resulting in skipping the check for release following
the LTS.

This was hardcoded to specific LTS release, so this
patch allows to match any SPDK release with LTS tag.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Change-Id: Ie85ab864114479f4e0b9b107796140249d1719fb
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/18443
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Ben Walker <benjamin.walker@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>