scripts: Add a script for backport handling.
Upload a script to be used for backporting
to other, maintained SPDK branches.

The script clones SPDK repository and uses information
about Gerrit user and hashtag to apply patches that are
present on `master` branch with the hashtag, but are not
yet on selected (hashtag) branch.
Additionally, this script will modify commit message of
these patches by:
 - deleting Reviewed-by, Tested-by lines,
 - adding the `(master)` to Reviewed-on link,
 - adding `(cherry picked from commit <original_commit_hash>)` line,
 - adding Gerrit user's signoff.

The script relies on uniqueness of Gerrit's Change-Id and
identifies patches based on this hash. If a backport has
already been made on desired branch, but its Change-Id
was different from original change on master branch,
then the script will include such patch.

Change-Id: Ic32b562d6d452f0de522bf784d733f1e8f7ac5af
Signed-off-by: Krzysztof Karas <krzysztof.karas@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/18704
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Jim Harris <jim.harris@gmail.com>