bdev_nvme: added max_bdevs and reported_bdevs in nvme_async_probe_ctx
Use max_bdevs to store the value that RPC bdev_nvme_attach_controller
requested and reported_bdevs to store the value to pass to the RPC
callback.
Removed the 'count' parameter in function populate_namespaces_cb().

This is a preparation for the following standby controller patches.

Change-Id: I219a06b79f15d230288477db28952ec26a25fd6b
Signed-off-by: Richael Zhuang <richael.zhuang@arm.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/18891
Reviewed-by: Jim Harris <jim.harris@gmail.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Community-CI: Mellanox Build Bot