thread: print error message if device I/O channel was not created
This is mostly relevant for the group channels that are created on the
modules themselves. If the error was not logged in that module, the subsequent call:

ch->group_ch = spdk_io_channel_get_ctx(spdk_get_io_channel(&module_if));

returns an invalid address (not NULL), which might be tricky to debug.

Signed-off-by: Denis Barakhtanov <denis.barahtanov@croit.io>
Change-Id: I0330a40254e70c60346aed13ea2b4c2520f28a72
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/18574
Community-CI: Broadcom CI <spdk-ci.pdl@broadcom.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>