nvme: remove remaining __attribute__((packed))
These strucutres had the packed attribute to make sure their size
changed whenever a new field was added at the end for ABI compatibility.
However, all of them have a corresponding *_get_opts() function, which
will will out the new fields, even if the code using it was unaware of
them when it was compiled.  For the same reason, their size can be
obtained with a regular sizeof(), and not SPDK_SIZEOF().

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I658ce6619ed92015b6c61d8460abd192be368496
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/18532
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Ben Walker <benjamin.walker@intel.com>